MachineLearning.Model.Measure

Description

Performance measures for Regression and Classification.

Documentation

data Measure

A performance measure has a string name and a function that takes a vector of the true values, a vector of predict values and returns a Double.

Constructors

Measure

Fields

_name :: String

_fun :: Vector -> Vector -> Double

true values -> predicted values -> measure

toMeasure :: String -> Measure

Read a string into a measure

measureAll :: [Measure]

List of all measures

_rmse :: Measure