MachineLearning.Model.Regression

Description

Fitting functions for the coefficients.

Documentation

fitTask :: Task -> TIR -> Dataset Double -> Vector Double -> [Vector Double]

chooses the appropriate fitting function fitTask :: Task -> Matrix Double -> Vector Double -> [Vector Double]

predictTask :: Task -> [Vector Double] -> Vector Double

chooses the appropriate prediction function

evalPenalty :: Penalty -> Int -> Double -> Double

evals the penalty function

applyMeasures :: [Measure] -> Vector Double -> Vector Double -> [Double]

applies a list of performance measures

nonlinearFit :: Int -> Matrix Double -> Matrix Double -> Vector Double -> (Vector Double -> Vector Double) -> (Vector Double -> Vector Double) -> Vector Double -> Vector Double

Non-linear optimization using Levenberg-Marquardt method. nonlinearFit :: Monad m => Vector Double -> Matrix Double -> Matrix Double -> Vector Double -> m (Vector Double)

tirToMatrix :: Dataset Double -> TIR -> (Matrix Double, Matrix Double)

transform a data matrix using a TIR expression. This function returns a tuple with the transformed data of the numerator and denominator, respectivelly. Each column of the transformed data represents one term of the TIR expression.