MachineLearning.TIR

Description

The TIR expression represents a function of the form:

\[ f(x) = g(\sum_{i}{w_i \cdot t_i(\prod_{j}{x_j^{k_{ij}})}} / (1 + \sum_{i}{w_i \cdot t_i(\prod_{j}{x_j^{k_{ij}})}}) \]

with \(t_i\) being a transformation function, \(g\) an invertible function, \(w_i\) a linear coefficient, \(k_{ij}\) the interaction strength.

Any given expression can be represented by two lists of terms, with each term being composed of a transformatioon function and an interaction. The transformation function is represented by a Function sum type. The interaction is represented as a list of tuple of ints where the key is the predictor index and the value is the strength of the predictor in this term. Strengths with a value of zero are omitted.

Documentation

data TIR

TIR is a record type composed of the external function _funY of type Function, numerator _p and denominator _q of type Sigma

Constructors

TIR

Fields

_funY :: Function

_p :: Sigma

_q :: Sigma

type Sigma = [Pi]

Sigma is just a list of terms Pi

type Pi = (Double, Function, [(Int, Int)])

Pi is a triple composed of a coefficient, a Function and a list of tuples where `(ix, k)` represents `x ! ix ^ k`.

randomRng :: (Int, Int) -> Rnd Int

generates a random integer within the specified range.

randomRngNZ :: (Int, Int) -> Rnd Int

generates a random integer within the specified range excluding zero.

randomFrom :: [a] -> Rnd a

picks a random element from a list.

randomVar :: MutationCfg -> Rnd (Maybe Int, MutationCfg)

returns a random index of variables provided by the mutation configuration.

randomVars :: MutationCfg -> Rnd [(Int, Int)]

returns a list of random interactions (tuples of variables indeces and exponentes) with parameters provided by the mutation configuration.

randomPi :: MutationCfg -> Rnd (Maybe Pi)

returns a random Pi

randomSigma :: MutationCfg -> Int -> Rnd (Sigma, Int)

returns a random Sigma

randomTIR :: MutationCfg -> Rnd TIR

returns a random TIR expression

type Column a = Vector a

We store thee dataset as a vector of columns. Each vector is stored a Storable-based vector.

type Dataset a = Vector (Column a)

A dataset is a Vector of Column

type Constraint = SRTree Int Double -> Double

A constraint is a function that gets a symbolic tree as an input and returns non negative Double representing how much a constraint was violated.

data Individual

An individual in the population is composed of the chromossome, a vector of fitness, a list of coefficients (for multiclass problems it stores one vector of coefficient per class), the constraint violation, the size of the expression, and the penalty value.

Constructors

Individual

Fields

_chromo :: TIR

_fit :: [Double]

_weights :: [Vector Double]

_constr :: Double

_len :: Int

_penalty :: Double

createIndividual :: TIR -> Individual

creates an unevaluated individual.

penalizedFit :: Individual -> Double

calculates the penalized fitness.

replaceConsts :: TIR -> Vector Double -> TIR

replaces the coefficients of a TIR expression

replaceWeight :: Pi -> State [Double] Pi

assembleTree :: Double -> TIR -> SRTree Int Double

creates a symbolic tree from a TIR expression.

prettyPrintsolution :: Individual -> String

pretty print a solution.