MachineLearning.Utils.Report

Description

Report generation.

Documentation

createIfDoesNotExist :: FilePath -> IO Handle

Creates a file if it does not exist

writeChampionStats :: Config -> (Individual -> Maybe [Double]) -> Int64 -> Individual -> IO ()

writes all the stats from the final champion solution

writeStats :: FilePath -> (Individual -> Maybe [Double]) -> [Measure] -> Int64 -> Individual -> IO ()

writes the stats to a file

writeCfg :: FilePath -> Config -> IO ()

creates a report file

writeExprs :: Task -> FilePath -> Individual -> IO ()

pretty write the expressions

evoLog :: Handle -> (Individual -> Maybe [Double]) -> Population Individual -> IO ()

creates a log of the evolution process

getStats :: [Double] -> [Double]

statsToStr :: Show a => [a] -> String

replaceWithNan :: Maybe [Double] -> Double

openNext :: String -> IO Handle

Opens the first file available in the format "name.{i}.csv" where i follows a sequence from 0 onward.

postAgg :: [Double] -> [Double]

aggregate :: [Double] -> Double -> [Double]

aggregates the best, worst and average solutions

createLogName :: Config -> FilePath

creates a log name instead of overwritting last one

makeLogger :: Config -> (Individual -> Maybe [Double]) -> IO (Population Individual -> IO (), Maybe Handle)

creates an IO logger function

closeIfJust :: Maybe Handle -> IO ()

closes a file